Update arbitration logic to prioritize high surplus and strict budget
